requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
pillow
ultralytics
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
